[aleph] Update for copying UWP and applying win32 icons

+++ bin/aleph/src/commands/build.rs
@@ -38,7 +38,7 @@ use aleph_target::build::target_platform;
 use aleph_target::Profile;
 use anyhow::anyhow;
 use clap::{Arg, ArgMatches};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::process::Command;
 
 pub struct Build {}
@@ -180,49 +180,10 @@ impl Build {
         self.plain_cargo_build(project, project_root, profile, target)
     }
 
-    fn plain_cargo_build(
-        &self,
-        project: ProjectSchema,
-        project_root: PathBuf,
-        profile: Profile,
-        target: &Target,
-    ) -> anyhow::Result<()> {
-        let mut command = match profile {
-            Profile::Debug => base_native_build(&project.game.crate_name)?,
-            Profile::Release => release_native_build(&project.game.crate_name)?,
-            Profile::Retail => retail_native_build(&project.game.crate_name)?,
-        };
-
-        self.add_win32_branding_env_vars(project, project_root, target, &mut command)?;
-
-        log::info!("{:?}", &command);
-        command.status()?;
-        Ok(())
-    }
-
-    fn add_win32_branding_env_vars(
-        &self,
-        project: ProjectSchema,
-        project_root: PathBuf,
-        target: &Target,
-        command: &mut Command,
-    ) -> anyhow::Result<()> {
-        // For windows or uwp we need to add an environment var for configuring the executable icon
-        if matches!(target.platform, BuildPlatform::Windows | BuildPlatform::UWP) {
-            if let Some(branding) = project.windows.and_then(|v| v.branding) {
-                let icon = branding.icon.as_ref();
-                let icon = resolve_absolute_or_root_relative_path(&project_root, icon);
-                log::info!("ALEPH_WIN32_ICON_FILE = {:#?}", &icon);
-                command.env("ALEPH_WIN32_ICON_FILE", icon);
-            }
-        }
-        Ok(())
-    }
-
     fn uwp(
         &self,
         project: ProjectSchema,
-        _project_root: PathBuf,
+        project_root: PathBuf,
         profile: Profile,
         target: &Target,
     ) -> anyhow::Result<()> {
@@ -231,8 +192,19 @@ impl Build {
             Profile::Release => release_uwp_build(target, &project.game.crate_name)?,
             Profile::Retail => retail_uwp_build(target, &project.game.crate_name)?,
         };
+
+        self.add_win32_branding_env_vars(&project, &project_root, target, &mut command)?;
+
         log::info!("{:?}", &command);
-        command.status()?;
+        let status = command.status()?;
+
+        if !status.success() {
+            log::error!("Cargo invocation failed! Terminating build.");
+            return Err(anyhow!("cargo invocation failed!"));
+        }
+
+        self.copy_uwp_build_to_appx_folder(&project, &project_root, profile, target)?;
+
         Ok(())
     }
 
@@ -249,8 +221,72 @@ impl Build {
             Profile::Retail => retail_android_build(target, &project.game.crate_name)?,
         };
         log::info!("{:?}", &command);
-        command.status()?;
+        let status = command.status()?;
 
+        if !status.success() {
+            log::error!("Cargo invocation failed! Terminating build.");
+            return Err(anyhow!("cargo invocation failed!"));
+        }
+
+        self.copy_android_build_to_gradle_project(project, project_root, profile, target)?;
+
+        Ok(())
+    }
+}
+
+impl Build {
+    fn plain_cargo_build(
+        &self,
+        project: ProjectSchema,
+        project_root: PathBuf,
+        profile: Profile,
+        target: &Target,
+    ) -> anyhow::Result<()> {
+        let mut command = match profile {
+            Profile::Debug => base_native_build(&project.game.crate_name)?,
+            Profile::Release => release_native_build(&project.game.crate_name)?,
+            Profile::Retail => retail_native_build(&project.game.crate_name)?,
+        };
+
+        self.add_win32_branding_env_vars(&project, &project_root, target, &mut command)?;
+
+        log::info!("{:?}", &command);
+        let status = command.status()?;
+
+        if !status.success() {
+            log::error!("Cargo invocation failed! Terminating build.");
+            return Err(anyhow!("cargo invocation failed!"));
+        }
+
+        Ok(())
+    }
+
+    fn add_win32_branding_env_vars(
+        &self,
+        project: &ProjectSchema,
+        project_root: &Path,
+        target: &Target,
+        command: &mut Command,
+    ) -> anyhow::Result<()> {
+        // For windows or uwp we need to add an environment var for configuring the executable icon
+        if matches!(target.platform, BuildPlatform::Windows | BuildPlatform::UWP) {
+            if let Some(branding) = project.windows.as_ref().and_then(|v| v.branding.as_ref()) {
+                let icon = branding.icon.as_ref();
+                let icon = resolve_absolute_or_root_relative_path(project_root, icon);
+                log::info!("ALEPH_WIN32_ICON_FILE = {:#?}", &icon);
+                command.env("ALEPH_WIN32_ICON_FILE", icon);
+            }
+        }
+        Ok(())
+    }
+
+    fn copy_android_build_to_gradle_project(
+        &self,
+        project: ProjectSchema,
+        project_root: PathBuf,
+        profile: Profile,
+        target: &Target,
+    ) -> anyhow::Result<()> {
         let android_project_root = target_project_root(target)?;
         if android_project_root.exists() {
             let mut output_dir = android_project_root.join("app");
@@ -278,6 +314,45 @@ impl Build {
                 if item.metadata()?.is_file() {
                     let item_path = item.path();
                     let dst = output_dir.join(item_path.file_name().unwrap());
+
+                    log::trace!("Copying '{:?}' -> '{:?}'", &item_path, &dst);
+                    std::fs::copy(item_path, dst)?;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    fn copy_uwp_build_to_appx_folder(
+        &self,
+        project: &ProjectSchema,
+        project_root: &PathBuf,
+        profile: Profile,
+        target: &Target,
+    ) -> anyhow::Result<()> {
+        let uwp_project_root = target_project_root(target)?;
+        if uwp_project_root.exists() {
+            let mut target_dir = project_root.join("target");
+            target_dir.push(format!("{}-uwp-windows-msvc", target.arch.name()));
+            target_dir.push(profile.name());
+
+            let cargo_metadata = get_cargo_metadata()?;
+            let (package, _) =
+                find_crate_and_target(&cargo_metadata, &project.game.crate_name, Some("cdylib"))?;
+            let exe_name = format!("{}.exe", &package.name);
+
+            let src_exe = target_dir.join(&exe_name);
+            let dst_exe = uwp_project_root.join(&exe_name);
+            log::trace!("Copying '{:?}' -> '{:?}'", src_exe, dst_exe);
+            std::fs::copy(src_exe, dst_exe)?;
+
+            let target_artifacts_dir = target_dir.join("artifacts");
+            for item in target_artifacts_dir.read_dir()? {
+                let item = item?;
+                if item.metadata()?.is_file() {
+                    let item_path = item.path();
+                    let dst = uwp_project_root.join(item_path.file_name().unwrap());
 
                     log::trace!("Copying '{:?}' -> '{:?}'", &item_path, &dst);
                     std::fs::copy(item_path, dst)?;
